api/types/system: move DiskUsage, DiskUsageOptions to api/types/backend
These types were introduced in f07242f6d7091fe9000aa0bea0cc27b85baa945a,
but while their description mentions it's the type used for the
response, it actually isn't, and it's used by the backend, but
ultimately marshaled to the "types.DiskUsage" struct;

https://github.com/moby/moby/blob/7dc46c6e0c727a746ecfb8427c0c3cbd998059bc/daemon/server/router/system/system_routes.go#L254-L270

(cherry picked from commit 82c069c857236edda446a901484b215fd82f2337)
